git clone source on the host instead of in the Dockerfile
This patch changes the way the source code is obtained. Instead of
performing a git clone in the Dockerfile, this steps is now done
one the host, and added to the image with a COPY.

If existing directories are specified for the runc and/or containerd
source through `RUNC_DIR` or `CONTAINERD_DIR`, that code is copied
to the same location as where we would otherwise clone the source code.

There's a number of reasons for doing this:

- Previously, the source code was always cloned in the Dockerfile,
  even if local source was used (`RUNC_DIR` or `CONTAINERD_DIR` set)
- There's no need to pass secrets to the `docker build` in case the
  source repository is a private repository. This makes it easier
  to use.
- source code is always obtained from the same location, and because
  a copy is used, there is no risk of modifying the original source
  code.
- Don't obtain the "master" git commit from GitHub; doing so made it
  more complicated to build from local source and/or from a fork.
  CI already does a "make clean" before building, so stale git
  repositories should not be present.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>