Makefile: allow overriding repository to build from
Our internal pipeline has a configuration to override the
repository to build from. While it's not currently used,
there could be situations where it's usefull (e.g. when
building an embargoed security release).

This patch adds a CONTAINERD_REMOTE makefile variable
that allows overriding the repository that we fetch the
source from.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>