Use upstream systemd unit file
Instead of maintaining our own systemd.unit file and keeping it in
sync with upstream, use the upstream one.

The upstream systemd unit assumes containerd is installed in /usr/local/bin,
whereas packages built using this repository are in stalled in /usr/bin, so
create a copy of the original unit, and replace the path before building.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>