Fix detection of default runc version due upstream to switch to go.mod
The code that was used to automatically select the default version of
runc to use for containerd was still assuming the containerd code to
have a vendor.conf.

Containerd "master", and v1.5 have switched to go modules, so should
now look in the go.mod file to get the recommended / default version.

Also change the order in which we check out source-code, so that the
containerd source-code is checked out before we determine the runc
version; this allows us to always use the local source of containerd,
without having to use curl to get the version from GitHub.

Finally, a small modification was made for the "make validate" target,
which now uses a generic "alpine" Golang image, so that validation can
be done before the containerd source code was checked out (which was
used to detect the Go version to use).

Before this patch:

    $ make docker.io/library/ubuntu:focal
    ...
    curl: (22) The requested URL returned error: 404 Not Found
    --------------------------------------------------------------------
    Building packages on docker.io/library/ubuntu:focal

    containerd   : HEAD (commit: a72fe7d)
    runc         :  (commit: 59ad417)
    architecture : x86_64
    build image  : docker.io/library/ubuntu:focal
    golang image : docker.io/library/golang:1.15.8-buster

With this patch:

    $ make docker.io/library/ubuntu:focal
    INFO: detected runc version from script/setup/runc-version
    ...
    Building packages on docker.io/library/ubuntu:focal

    containerd   : HEAD (commit: a72fe7d)
    runc         : v1.0.0-rc93 (commit: 59ad417)
    architecture : x86_64
    build image  : docker.io/library/ubuntu:focal
    golang image : docker.io/library/golang:1.15.8-buster

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>