Use OS native runc rather than bundling on RHEL8
The RPM does provide on runc for most platforms, but not for RHEL8. It
does not work there, since there is native package that provides runc
that lives in a module stream. Yum fails to find or install a non-modular
package when a module provides the same name.

For those with access to RedHat pages, some more information is available
here: https://access.redhat.com/solutions/5588101

A problem with simply omitting the "provides" from the package is that it
will then come in conflict with the native package. It is for example not
possible to install both docker and the native podman in parallel.

This patch changes the contanerd.io package on RHEL8 to require the native
OS version instead of bundling its own copy. With this, parallel installation
of docker and podman is again possible.