Fix RUNC_REF not being evaluated correctly
When testing the script locally from a clean state, I noticed that the RUNC_REF
variable was not working as expected, and failed to pick the runc version that
is set in containerd.

This patch fixes the runc version not being detected correctly.

While Make "promises" to expand variables "when used", it actually performs
this earlier than that, so when combining, it evaluates $(SOME_VAR) *before*
the target is run....

From the docs: https://www.gnu.org/software/make/manual/html_node/Flavors.html#Flavors

> references are expanded whenever this variable is substituted (in the course
> of expanding some other string). When this happens, it is called recursive expansion.

However, taking the following Makefile:

    SOME_VAR=$(shell cat hello)

    .PHONY: clean
    clean:
        @rm hello

    .PHONY: together
    together:
        @echo this is hello > hello
        @cat hello
        @echo SOME_VAR is \""$(SOME_VAR)"\"

    .PHONY: separate
    separate: one two

    .PHONY: one
    one:
        @echo this is hello > hello
        @cat hello

    .PHONE: two
    two:
        @echo SOME_VAR is \""$(SOME_VAR)"\"

When combined:

    make clean
    make together
    cat: hello: No such file or directory
    this is hello
    SOME_VAR is ""

As shown, $(SOME_VAR) is evaluated *before* `together` is run, and produces
an error, because "hello" is not yet created.

When doing creating the file and using the file separate, things work as
expected, and $(SOME_VAR) is able to read the file:

    make clean
    make separate
    this is hello
    SOME_VAR is "this is hello"

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>