Fix missing echo in scripts/determine-runc-version
This was introduced in e1b2c4a0e7437e63b018829deb4d0e79e8065205, where
the version that was found is now stored in a variable (runc_ref) for
debugging.

However, I forgot to add an "echo" to actually print the version.

Because of that, the script debugs the version it found:

    + git -C src/github.com/containerd/containerd checkout -q refs/tags/v1.4.6
    ./scripts/checkout.sh src/github.com/opencontainers/runc "$(./scripts/determine-runc-version)"
    INFO: detected runc version (v1.0.0-rc95) from script/setup/runc-version

But in the "build" target it shows that it didn't actually print it
(note the missing version missing between `runc :` and `(commit: ...)`):

    --------------------------------------------------------------------
    Building packages on docker.io/dockereng/rhel:7-s390x

    containerd   : v1.4.6 (commit: d71fcd7)
    INFO: detected runc version (v1.0.0-rc95) from script/setup/runc-version
    runc         :  (commit: e005fee)
    architecture : s390x
    build image  : docker.io/dockereng/rhel:7-s390x
    golang image : docker.io/library/golang:1.13.15-buster
    --------------------------------------------------------------------

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>