Makefile: allow GOVERSION to be passed, but being empty
Some of our Jenkins jobs allow for GOVERSION to be set as parameter but
unconditionally pass the variable when calling `make` from this repository.
The intent of those parameters was to allow overriding the Go version when
building packages, but due to the parameter to be passed unconditionally,
curently requires us to always set the Go version to build, and making sure
it's set to the correct version for containerd to build.

Make (unfortunately) seems to make it nigh impossible to detect if a variable
is either unset (not passed) or empty, while also accounting for the variable
to be passed as `FOO= make <target>` and `make FOO= <target>` _and_ allowing
the default value to be evaluated lazily (on "use"), which is needed in our
case (to get the default version from containerd's repository, which requires
the source code to be checked out first).

Variations of the below all failed (as demonstrated below);

```makefile
ifndef FOO
	FOO = $(shell echo "default1")
else ifeq ($(strip $(FOO)),)
	FOO = $(shell echo "default2")
endif

.PHONY: test
test:
	@echo foo is "$(FOO)" from $(origin FOO)
```

```bash
$ make test
foo is default1 from file

$ FOO= make test
foo is default1 from file

$ make FOO= test
foo is  from command line
```

To work around this, we're introducing separate variables for the
`GOVERSION` (to be used to override the default) and `GOLANG_VERSION`,
which is the (lazily evaluated) variable used to construct the golang
image reference to use.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>