scripts: make sure all dates use UTC
While most of the dates generated were done from within a container (which defaults
to using UTC), some scripts could be run on the host, and therefore depend on the
host's configured timezone.

This patch makes sure that all calls to `date` use `TZ=UTC` to make sure the dates
are consistent.

Before this patch:

    ./scripts/new-deb-release 1.2.3 &> /dev/null && git diff -- debian/changelog | grep Sebastiaan
    + -- Sebastiaan van Stijn <github@gone.nl>  Wed, 06 Apr 2022 12:22:09 +0200

With this patch:

    ./scripts/new-deb-release 1.2.3 &> /dev/null && git diff -- debian/changelog | grep Sebastiaan
    + -- Sebastiaan van Stijn <github@gone.nl>  Wed, 06 Apr 2022 10:22:15 +0000

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>