Jenkinsfile: remove Debian and Raspbian 10 Buster (EOL LTS: 2024-06-30)
Remove this version of Debian, which reaches end of the line. End of standard
support already reached EOL on 2022-09-10, and stopped advertising this version
of Debian in our documentation, but builds were still running.

This patch removes it from CI in this repository (but could still be built
in our release pipeline).

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>