Jenkinsfile: remove CentOS 7 (EOL 2024-06-30)
CentOS 7 [reaches EOL on June 30][1], so we won't be building packages
for containerd 1.7. containerd 1.6 packages are still available for the
remaining time for those that need it.

This patch removes it from CI in this repository (but could still be built
in our release pipeline).

[1]: https://blog.centos.org/2023/04/end-dates-are-coming-for-centos-stream-8-and-centos-linux-7/#centos-linux-7-end-of-life-june-30-2024

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>
(cherry picked from commit 6bc9688c86b6fbd1d3b6cdceed44e28756db53aa)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>