dockerfiles/deb: Install binutils-gold on arm64 if needed
On Debian Trixie and newer versions, binutils-gold is no longer included
in the base binutils package and must be installed separately. Without
this component, the build fails on arm64 with "collect2: fatal error:
cannot find 'ld'" error.

Go has a hardcoded workaround where it forces usage of ld.gold on ARM64.
It was only needed for ld 2.40, and Trixie is already on 2.44 but the
workaround hasn't been removed from the upstream Go code yet.

We need to make sure that ld.gold is available when building for Trixie
on arm64.

This is the same case as with Fedora 41: b1534837b42f81528c4228e6518dee0df0662583

Signed-off-by: Paweł Gronowski <pawel.gronowski@docker.com>