Use SOURCE_DATE_EPOCH to improve reproducibility
Set `SOURCE_DATE_EPOCH` to the most recent modification timestamp of all
repos under `src/`. This allows deb and rpm toolchains to clamp mtimes
and better reproduce packages that would otherwise be different due
solely to the current time.

See https://reproducible-builds.org/docs/source-date-epoch/ for a
recommended approaches to setting this variable and a list of supported
tooling. Support varies between versions of rpm deb toolchains so
packages may not be bit-by-bit reproducible in all cases.

Note that this was developed while testing a fix to #303 but it does not
address that issue directly as compiled binaries still differ between
Debian releases.

Signed-off-by: Dan Duvall <dduvall@wikimedia.org>